+++ src/components/modules/escrow/server/escrow-firebase.ts
@@ -16,7 +16,7 @@ import {
   where,
 } from "firebase/firestore";
 import { z } from "zod";
-import { formSchema } from "../schema/initialize-escrow-schema";
+import { formSchema } from "../schema/initialize-escrow.schema";
 
 interface addEscrowProps {
   payload: z.infer<typeof formSchema>;
